refactor(logic): Main 클래스 추가 클래스명 변경
