+++ src/main/java/com/rick/racing/CarRacing.java
@@ -1,47 +1,11 @@
 package com.rick.racing;
 
-import com.rick.racing.model.CarRecordHistory;
-import com.rick.racing.model.RacingPlayData;
-import com.rick.racing.model.RacingResult;
-import com.rick.racing.ui.InputView;
-import com.rick.racing.ui.ResultView;
-import java.util.Random;
-import java.util.function.Supplier;
+import com.rick.racing.core.CarRacingCore;
 
 public class CarRacing {
 
-    public static final int RANDOM_BOUND = 10;
-    public static final int GO_THRESHOLD = 4;
-    private final InputView inputView;
-    private final ResultView resultView;
-    private final Random random;
-
-    public CarRacing() {
-        inputView = new InputView();
-        resultView = new ResultView();
-        random = new Random();
+    public static void main(String[] args) {
+        new CarRacingCore().start();
     }
 
-    public void start() {
-        RacingPlayData racingPlayData = inputView.getGameDataFromUser();
-        resultView.showResult(racingPlayData, doGame(racingPlayData, this::isGo));
-    }
-
-    public RacingResult doGame(RacingPlayData racingPlayData, Supplier<Boolean> goOrStopSupplier) {
-        RacingResult racingResult = new RacingResult();
-
-        for (int carIndex = 0; carIndex < racingPlayData.getCarCount(); carIndex++) {
-            CarRecordHistory carRecordHistory = new CarRecordHistory();
-            for (int tryIndex = 0; tryIndex < racingPlayData.getTryCount(); tryIndex++) {
-                carRecordHistory.addHistory(goOrStopSupplier.get());
-            }
-            racingResult.addRecord(carRecordHistory);
-
-        }
-        return racingResult;
-    }
-
-    private boolean isGo() {
-        return random.nextInt(RANDOM_BOUND) >= GO_THRESHOLD;
-    }
 }
